// React/app/GerAdv_TS/GUTAtividades/Interfaces/interface.GUTAtividades.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface IGUTAtividades {
// 202501251
    id: number;
 
	gutperiodicidade: number,
	operador: number,
	nome: string,
	observacao: string,
	gutgrupo: number,
	concluido: boolean,
	dataconcluido: string,
	diasparainiciar: number,
	minutospararealizar: number,
}

export interface IGUTAtividadesFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IGUTAtividadesIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}